Revert "Use yarn when running inside yarn workspace. (#3997)"
This reverts commit 2c34d5b66eab7d1c96e573bc48b8e82b6d8e82b0.
