+++ packages/create-react-app/package.json
@@ -27,7 +27,6 @@
     "envinfo": "5.4.0",
     "fs-extra": "^5.0.0",
     "hyperquest": "^2.1.2",
-    "react-dev-utils": "^5.0.0",
     "semver": "^5.0.3",
     "tar-pack": "^3.4.0",
     "tmp": "0.0.33",
