+++ packages/react-dev-utils/package.json
@@ -34,8 +34,7 @@
     "printHostingInstructions.js",
     "WatchMissingNodeModulesPlugin.js",
     "WebpackDevServerUtils.js",
-    "webpackHotDevClient.js",
-    "workspaceUtils.js"
+    "webpackHotDevClient.js"
   ],
   "dependencies": {
     "@babel/code-frame": "7.0.0-beta.46",
@@ -46,9 +45,7 @@
     "detect-port-alt": "1.1.6",
     "escape-string-regexp": "1.0.5",
     "filesize": "3.6.1",
-    "find-pkg": "1.0.0",
     "global-modules": "1.0.0",
-    "globby": "8.0.1",
     "gzip-size": "4.1.0",
     "inquirer": "5.1.0",
     "is-root": "1.0.0",
