+++ packages/react-scripts/package.json
@@ -44,7 +44,9 @@
     "eslint-plugin-jsx-a11y": "6.0.3",
     "eslint-plugin-react": "7.8.2",
     "file-loader": "1.1.11",
+    "find-pkg": "1.0.0",
     "fs-extra": "5.0.0",
+    "globby": "7.1.1",
     "graphql": "0.13.2",
     "graphql-tag": "2.9.2",
     "html-webpack-plugin": "3.2.0",
