+++ packages/react-scripts/scripts/build.js
@@ -46,6 +46,7 @@ const { printBrowsers } = require('react-dev-utils/browsersHelper');
 const measureFileSizesBeforeBuild =
   FileSizeReporter.measureFileSizesBeforeBuild;
 const printFileSizesAfterBuild = FileSizeReporter.printFileSizesAfterBuild;
+const useYarn = fs.existsSync(paths.yarnLockFile);
 
 // These sizes are pretty large. We'll warn for bundles exceeding them.
 const WARN_AFTER_BUNDLE_GZIP_SIZE = 512 * 1024;
@@ -116,7 +117,7 @@ checkBrowsers(paths.appPath)
         publicUrl,
         publicPath,
         buildFolder,
-        paths.useYarn
+        useYarn
       );
       printBrowsers(paths.appPath);
     },
